Isolate test case, testExcessiveCheckpointingDuringPushReplication(), from ReplicationTest.java
It seems other test cases impact to the result of test case, testExcessiveCheckpointingDuringPushReplication().
I guess that static variable changes of java core classes for unit test purpose might cause this problem.
Instead of reviewing all unit test cases in ReplicationTest.java, I simply extract testExcessiveCheckpointingDuringPushReplication() from ReplicationTest.java.
